refactor: separate crossref and doi utils (#86)
