+++ __tests__/api-atlases-id-source-datasets-create.test.ts
@@ -20,7 +20,7 @@ import {
 import { TestAtlas, TestUser } from "../testing/entities";
 
 jest.mock("../app/utils/pg-app-connect-config");
-jest.mock("../app/utils/publications");
+jest.mock("../app/utils/crossref");
 jest.mock("../app/utils/hca-projects");
 
 const NEW_DATASET_DATA: NewSourceDatasetData = {
